HUGE UPDATE!
massive system overhaul and system optimisations. UI updated, Time step system implemented, some UI elements currently disabled due to FPS impact on use with large files.

+++ Assets/Scripts/OrbitalData.cs
@@ -43,7 +43,7 @@
         public Display Display;
 
         [JsonProperty("radius")]
-        public long Radius;
+        public double Radius;
 
         [JsonProperty("eph")]        
         //public Eph[] eph;
